feat: Implementation of user registration
Added controllers, validation schemes and entities for user registration via email and password.

Implemented basic registration logic with checking for the existence of a user with that email and storing the hashed password.

Added paths for registration and authorization via Google.

Configured project infrastructure, including tsconfig.json to support paths and types, and a global variable to access DataSource.