feat: Implementation of endpoint for getting a questions (master)
Added new endpoint `/v1/batch` for getting a package of questions quiz based on a text fragment.

Implemented logic of request to Gemini API for generating questions based on the provided text. Added validation schemes of incoming parameters and Gemini response.

Created a mechanism for reading a random paragraph from a text file for use as source text for generating questions. to be used as source text for question generation.