feat: Added description to questions (master)
Added `description` column to the `Question` entity to store the description of the question. This will allow users to better understand the context of the question and explain why a particular answer is correct.  Updated the OpenAPI schema for the `batchController` to include the field
`description` field in the query and response.